pyproject.toml
[tool.poetry]
name = "pykasso"
version = "0.1.0"
description = "Python project intended to simulate stochastic karst network"
license = "GPL-3.0"
authors = [
    "François Miville <francois@miville.org>",
    "Chloé Fandel",
    "Philippe Renard"
]
maintainers = [
    "François Miville <francois@miville.org>"
]
readme = "README.md"
homepage = "https://github.com/randlab/pyKasso"
repository = "https://github.com/randlab/pyKasso"
documentation = "" # TODO
keywords = ['hydrogeology', '3-D modeling', 'stochasticity']
classifiers = [
    "Development Status :: 3 - Alpha",
    "Programming Language :: Python :: 3.9",
    "Operating System :: OS Independent",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Natural Language :: English",
    "Topic :: Scientific/Engineering",
]

[tool.poetry.dependencies]
python = "^3.9"
numpy = "^1.22.0"
pandas = "^1.5.2"
matplotlib = "^3.6.2"
openpyxl = "^3.0.10"
mpmath = "^1.2.1"
pyyaml = "^6.0"
agd = "^0.1.31"
scipy = "^1.9.3"
plotly = "^5.13.1"
pytest = "^7.2.2"

[tool.poetry.group.analysis.dependencies]
karstnet = { git = "https://github.com/karstnet/karstnet.git", branch = "master" }
networkx = "^2.8.8"
mplstereonet = "^0.6.2"

[tool.poetry.group.visualization.dependencies]
pyvista = "^0.37.0"
imageio = "^2.26.1"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.19.4"
jupyter = "^1.0.0"
jupyterlab = "^3.5.2"
ipyvtklink = "^0.2.3"
snakeviz = "^2.1.1"
sphinx = "^6.1.3"
numpydoc = "^1.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"